feat: globe history slider, map sidebar, altitude legend
Three related follow-ups to the 3D globe and flat map:

- History mode's 1h/6h/24h buttons replaced with a range slider
  (15-minute/0.25h steps, up to 24h). GET /api/tracks's `hours` param
  changed from int (ge=1) to float (ge=0.25) to support it. Added a
  request-token guard in enterHistoryMode since a dragged slider (unlike
  discrete buttons) can fire several requests in quick succession and
  responses aren't guaranteed to resolve in order.
- Historical tracks now actually respect the track-opacity setting --
  they were being drawn at full opacity, never calling .withAlpha() the
  way live tracks already did.
- New shared altitude-legend.js renders a color-swatch legend from
  /api/config's altitude_bands (which already carries a human-readable
  label per band, previously discarded by both globe.js and map.js).
  Added to the 3D globe, the embedded dashboard map, and the fullscreen
  map.
- map.js's track-line click handler now also opens the same shared
  aircraft detail sidebar every other page uses (previously it only
  toggled a line-width highlight) -- lands on both index.html and
  fullmap.html since they share map.js.

Verified live: slider label formats correctly at 15-minute granularity,
switching to a history window draws visibly more transparent tracks,
legend renders on both pages, and clicking a track on the fullscreen map
opens the full sidebar. Zero console errors.

+++ app/db/queries/tracks.py
@@ -42,7 +42,7 @@ class AircraftTrack:
     segments: list[list[TrackPoint]]
 
 
-async def get_tracks(pool: asyncpg.Pool, hours: int) -> list[AircraftTrack]:
+async def get_tracks(pool: asyncpg.Pool, hours: float) -> list[AircraftTrack]:
     since = datetime.now(UTC) - timedelta(hours=hours)
 
     top_icaos_rows = await pool.fetch(
